+++ test/expression/evaluateStep.mocha.js
@@ -37,11 +37,11 @@ let diamonds = External.fromJS({
     { name: 'isNice', type: 'BOOLEAN' },
     { name: 'tags', type: 'SET/STRING' },
     { name: 'pugs', type: 'SET/STRING' },
-    { name: 'carat', type: 'NUMBER' },
+    { name: 'carat', type: 'NUMBER', nativeType: 'STRING' },
     { name: 'height_bucket', type: 'NUMBER' },
     { name: 'price', type: 'NUMBER', unsplitable: true },
     { name: 'tax', type: 'NUMBER', unsplitable: true },
-    { name: 'vendor_id', special: 'unique', unsplitable: true }
+    { name: 'vendor_id', type: 'NULL', nativeType: 'hyperUnique', unsplitable: true }
   ],
   allowEternity: true,
   allowSelectQueries: true