+++ test/expression/resolve.mocha.js
@@ -246,7 +246,7 @@ describe("resolve", () => {
           { name: '__time', type: 'TIME' },
           { name: 'color', type: 'STRING' },
           { name: 'cut', type: 'STRING' },
-          { name: 'carat', type: 'NUMBER' }
+          { name: 'carat', type: 'NUMBER', nativeType: 'STRING' }
         ]
       });
 
@@ -279,7 +279,7 @@ describe("resolve", () => {
         { name: '__time', type: 'TIME' },
         { name: 'color', type: 'STRING' },
         { name: 'cut', type: 'STRING' },
-        { name: 'carat', type: 'NUMBER' }
+        { name: 'carat', type: 'NUMBER', nativeType: 'STRING' }
       ]
     });